+++ runtime/test/unit/runner.clj
@@ -21,6 +21,7 @@
             [ujima.control.queries-test]
             [ujima.api-test]
             [ujima.api.routes-test]
+            [ujima.api.auth-test]
 
             [ujima.linux.audio-test]
             [ujima.linux.i3-test]
@@ -58,6 +59,7 @@
     ujima.control.queries-test
     ujima.api-test
     ujima.api.routes-test
+    ujima.api.auth-test
 
     ujima.linux.audio-test
     ujima.linux.i3-test